3. LLM Application with Retrieval Augmented Generate (RAG)/requirements.txt
faiss-cpu==1.11.0
numpy==1.26.4
openai==1.97.0
pandas==2.2.3
sentence-transformers==2.6.1
streamlit==1.35.0